Refactor dispatching into a separate object per RPCServer instance
A Dispatcher object is created for each RPCServer instance, allowing for
multiple clients to connect and execute methods without the possibility
of request ID conflicts in a global dispatcher.
