+++ src/rpcx/__init__.py
@@ -1,3 +1,2 @@
 from .client import RPCClient
-from .manager import RPCManager, Stream
-from .server import RPCServer
+from .server import RPCManager, RPCServer, Stream
